+++ junopoker/src/main/java/com/v1/junopoker/dto/EndPlayerActionRequest.java
@@ -3,15 +3,17 @@ package com.v1.junopoker.dto;
 import com.v1.junopoker.model.Card;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class EndPlayerActionRequest {
     private RequestType type;
     private char action;
     private String username;
     private int seat;
-    private double bet;
-    private double stackSize;
-    private double potSize;
-    private double currentStreetPotSize;
+    private BigDecimal bet;
+    private BigDecimal stackSize;
+    private BigDecimal potSize;
+    private BigDecimal currentStreetPotSize;
     private boolean isPreFlop;
 }
