+++ junopoker/src/main/java/com/v1/junopoker/dto/InitPotRequest.java
@@ -2,12 +2,14 @@ package com.v1.junopoker.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class InitPotRequest {
     private RequestType type;
     private int smallBlind;
     private int bigBlind;
-    private double sbAmount;
-    private double bbAmount;
-    private double potSize;
+    private BigDecimal sbAmount;
+    private BigDecimal bbAmount;
+    private BigDecimal potSize;
 }
